reformatted the project for consistency
Revert "partial refactor, quash later"

This reverts commit 73743527a0d772143ad59d21de687975d16ffaec.
a rename managedto break the build for some reason???

also a mini-commit, quash later.

renamed stuff to be more consistent

clangformated the entire project

missed some

missed some more

+++ stdlib/include/io.h
@@ -9,5 +9,5 @@ typedef struct str_view_t str_view;
 int selfputchar(unsigned char c);
 int selfputcharerr(unsigned char c);
 int selfprint(str_view s);
-int printmulti(unsigned char c, int n);
+int printmulti(unsigned char c, int64_t n);
 int geterrno();
