+++ stdlib/io.c
@@ -19,8 +19,8 @@ int selfprint(str_view s) {
   return 0;
 }
 
-int printmulti(unsigned char c, int n) {
-  for (int i = 0; i < n; i++) {
+int printmulti(unsigned char c, int64_t n) {
+  for (int64_t i = 0; i < n; i++) {
     int result = selfputchar(c);
     if (errno) {
       fputs(strerror(result), stderr);
